images/azure-ambari-prometheus-exporter/requirements.txt
PyYAML
prometheus_client
statsd
requests
